requirements/default.txt
###### Django ######
Django==1.8.2
djangorestframework==3.1.2
django-model-utils==2.5
django-extensions==1.6.3

#### AI #####
aiml==0.8.6